Struct generate_random::Client

pub(crate) struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> {
    handle: Arc<Handle<C, M, R>>,
}

An ergonomic service client for TrentService.

This client allows ergonomic access to a TrentService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Fields
handle: Arc<Handle<C, M, R>>
Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Client<C, M, R>

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn cancel_key_deletion(&self) -> CancelKeyDeletion<C, M, R>

Constructs a fluent builder for the CancelKeyDeletion operation.

See CancelKeyDeletion for more information about the operation and its arguments.

pub fn connect_custom_key_store(&self) -> ConnectCustomKeyStore<C, M, R>

Constructs a fluent builder for the ConnectCustomKeyStore operation.

See ConnectCustomKeyStore for more information about the operation and its arguments.

pub fn create_alias(&self) -> CreateAlias<C, M, R>

Constructs a fluent builder for the CreateAlias operation.

See CreateAlias for more information about the operation and its arguments.

pub fn create_custom_key_store(&self) -> CreateCustomKeyStore<C, M, R>

Constructs a fluent builder for the CreateCustomKeyStore operation.

See CreateCustomKeyStore for more information about the operation and its arguments.

pub fn create_grant(&self) -> CreateGrant<C, M, R>

Constructs a fluent builder for the CreateGrant operation.

See CreateGrant for more information about the operation and its arguments.

pub fn create_key(&self) -> CreateKey<C, M, R>

Constructs a fluent builder for the CreateKey operation.

See CreateKey for more information about the operation and its arguments.

pub fn decrypt(&self) -> Decrypt<C, M, R>

Constructs a fluent builder for the Decrypt operation.

See Decrypt for more information about the operation and its arguments.

pub fn delete_alias(&self) -> DeleteAlias<C, M, R>

Constructs a fluent builder for the DeleteAlias operation.

See DeleteAlias for more information about the operation and its arguments.

pub fn delete_custom_key_store(&self) -> DeleteCustomKeyStore<C, M, R>

Constructs a fluent builder for the DeleteCustomKeyStore operation.

See DeleteCustomKeyStore for more information about the operation and its arguments.

pub fn delete_imported_key_material(&self) -> DeleteImportedKeyMaterial<C, M, R>

Constructs a fluent builder for the DeleteImportedKeyMaterial operation.

See DeleteImportedKeyMaterial for more information about the operation and its arguments.

pub fn describe_custom_key_stores(&self) -> DescribeCustomKeyStores<C, M, R>

Constructs a fluent builder for the DescribeCustomKeyStores operation.

See DescribeCustomKeyStores for more information about the operation and its arguments.

pub fn describe_key(&self) -> DescribeKey<C, M, R>

Constructs a fluent builder for the DescribeKey operation.

See DescribeKey for more information about the operation and its arguments.

pub fn disable_key(&self) -> DisableKey<C, M, R>

Constructs a fluent builder for the DisableKey operation.

See DisableKey for more information about the operation and its arguments.

pub fn disable_key_rotation(&self) -> DisableKeyRotation<C, M, R>

Constructs a fluent builder for the DisableKeyRotation operation.

See DisableKeyRotation for more information about the operation and its arguments.

pub fn disconnect_custom_key_store(&self) -> DisconnectCustomKeyStore<C, M, R>

Constructs a fluent builder for the DisconnectCustomKeyStore operation.

See DisconnectCustomKeyStore for more information about the operation and its arguments.

pub fn enable_key(&self) -> EnableKey<C, M, R>

Constructs a fluent builder for the EnableKey operation.

See EnableKey for more information about the operation and its arguments.

pub fn enable_key_rotation(&self) -> EnableKeyRotation<C, M, R>

Constructs a fluent builder for the EnableKeyRotation operation.

See EnableKeyRotation for more information about the operation and its arguments.

pub fn encrypt(&self) -> Encrypt<C, M, R>

Constructs a fluent builder for the Encrypt operation.

See Encrypt for more information about the operation and its arguments.

pub fn generate_data_key(&self) -> GenerateDataKey<C, M, R>

Constructs a fluent builder for the GenerateDataKey operation.

See GenerateDataKey for more information about the operation and its arguments.

pub fn generate_data_key_pair(&self) -> GenerateDataKeyPair<C, M, R>

Constructs a fluent builder for the GenerateDataKeyPair operation.

See GenerateDataKeyPair for more information about the operation and its arguments.

pub fn generate_data_key_pair_without_plaintext(
&self
) -> GenerateDataKeyPairWithoutPlaintext<C, M, R>

Constructs a fluent builder for the GenerateDataKeyPairWithoutPlaintext operation.

See GenerateDataKeyPairWithoutPlaintext for more information about the operation and its arguments.

pub fn generate_data_key_without_plaintext(
&self
) -> GenerateDataKeyWithoutPlaintext<C, M, R>

Constructs a fluent builder for the GenerateDataKeyWithoutPlaintext operation.

See GenerateDataKeyWithoutPlaintext for more information about the operation and its arguments.

pub fn generate_random(&self) -> GenerateRandom<C, M, R>

Constructs a fluent builder for the GenerateRandom operation.

See GenerateRandom for more information about the operation and its arguments.

pub fn get_key_policy(&self) -> GetKeyPolicy<C, M, R>

Constructs a fluent builder for the GetKeyPolicy operation.

See GetKeyPolicy for more information about the operation and its arguments.

pub fn get_key_rotation_status(&self) -> GetKeyRotationStatus<C, M, R>

Constructs a fluent builder for the GetKeyRotationStatus operation.

See GetKeyRotationStatus for more information about the operation and its arguments.

pub fn get_parameters_for_import(&self) -> GetParametersForImport<C, M, R>

Constructs a fluent builder for the GetParametersForImport operation.

See GetParametersForImport for more information about the operation and its arguments.

pub fn get_public_key(&self) -> GetPublicKey<C, M, R>

Constructs a fluent builder for the GetPublicKey operation.

See GetPublicKey for more information about the operation and its arguments.

pub fn import_key_material(&self) -> ImportKeyMaterial<C, M, R>

Constructs a fluent builder for the ImportKeyMaterial operation.

See ImportKeyMaterial for more information about the operation and its arguments.

pub fn list_aliases(&self) -> ListAliases<C, M, R>

Constructs a fluent builder for the ListAliases operation.

See ListAliases for more information about the operation and its arguments.

pub fn list_grants(&self) -> ListGrants<C, M, R>

Constructs a fluent builder for the ListGrants operation.

See ListGrants for more information about the operation and its arguments.

pub fn list_key_policies(&self) -> ListKeyPolicies<C, M, R>

Constructs a fluent builder for the ListKeyPolicies operation.

See ListKeyPolicies for more information about the operation and its arguments.

pub fn list_keys(&self) -> ListKeys<C, M, R>

Constructs a fluent builder for the ListKeys operation.

See ListKeys for more information about the operation and its arguments.

pub fn list_resource_tags(&self) -> ListResourceTags<C, M, R>

Constructs a fluent builder for the ListResourceTags operation.

See ListResourceTags for more information about the operation and its arguments.

pub fn list_retirable_grants(&self) -> ListRetirableGrants<C, M, R>

Constructs a fluent builder for the ListRetirableGrants operation.

See ListRetirableGrants for more information about the operation and its arguments.

pub fn put_key_policy(&self) -> PutKeyPolicy<C, M, R>

Constructs a fluent builder for the PutKeyPolicy operation.

See PutKeyPolicy for more information about the operation and its arguments.

pub fn re_encrypt(&self) -> ReEncrypt<C, M, R>

Constructs a fluent builder for the ReEncrypt operation.

See ReEncrypt for more information about the operation and its arguments.

pub fn replicate_key(&self) -> ReplicateKey<C, M, R>

Constructs a fluent builder for the ReplicateKey operation.

See ReplicateKey for more information about the operation and its arguments.

pub fn retire_grant(&self) -> RetireGrant<C, M, R>

Constructs a fluent builder for the RetireGrant operation.

See RetireGrant for more information about the operation and its arguments.

pub fn revoke_grant(&self) -> RevokeGrant<C, M, R>

Constructs a fluent builder for the RevokeGrant operation.

See RevokeGrant for more information about the operation and its arguments.

pub fn schedule_key_deletion(&self) -> ScheduleKeyDeletion<C, M, R>

Constructs a fluent builder for the ScheduleKeyDeletion operation.

See ScheduleKeyDeletion for more information about the operation and its arguments.

pub fn sign(&self) -> Sign<C, M, R>

Constructs a fluent builder for the Sign operation.

See Sign for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

pub fn update_alias(&self) -> UpdateAlias<C, M, R>

Constructs a fluent builder for the UpdateAlias operation.

See UpdateAlias for more information about the operation and its arguments.

pub fn update_custom_key_store(&self) -> UpdateCustomKeyStore<C, M, R>

Constructs a fluent builder for the UpdateCustomKeyStore operation.

See UpdateCustomKeyStore for more information about the operation and its arguments.

pub fn update_key_description(&self) -> UpdateKeyDescription<C, M, R>

Constructs a fluent builder for the UpdateKeyDescription operation.

See UpdateKeyDescription for more information about the operation and its arguments.

pub fn update_primary_region(&self) -> UpdatePrimaryRegion<C, M, R>

Constructs a fluent builder for the UpdatePrimaryRegion operation.

See UpdatePrimaryRegion for more information about the operation and its arguments.

pub fn verify(&self) -> Verify<C, M, R>

Constructs a fluent builder for the Verify operation.

See Verify for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(
conf: Config,
conn: C
) -> Client<C, AwsMiddleware, Standard>

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Client<DynConnector, AwsMiddleware, Standard>

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Client<DynConnector, AwsMiddleware, Standard>

Creates a new client from the service Config.